requirements.txt
Flask==0.10.1
Jinja2==2.7.2
Werkzeug==0.9.4
gunicorn==19.3.0
Pillow==2.9.0
requests==2.7.0
